Endpoints that need protection now have API authentication

+++ kuky/routes/kus.js
@@ -2,6 +2,7 @@ var express = require('express');
 var router = express.Router();
 var models  = require('../models');
 var sequelize = require('sequelize');
+var apiAuth = require('./apiAuth');
 
 var User = models.sequelize.models.User;
 var Ku = models.sequelize.models.Ku;
@@ -18,7 +19,7 @@ var karmaThreshold = 3;
 /*
 GET single ku by id
 */
-router.get('/:id', function (req, res, next) {
+router.get('/:id', apiAuth.authenticate, function (req, res, next) { // TODO:FIX THIS
     var response = {
         ku: {},
         comments: []
@@ -39,15 +40,15 @@ router.get('/:id', function (req, res, next) {
                     response['comments'].push(elem.getData());
                 })
                 res.json(response);
-            })
-        })
-    })
-})
+            });
+        });
+    });
+});
 
 /* 
 GET all kus in new order
  */
-router.get('/all/:sort', function (req, res, next) {
+router.get('/all/:sort', apiAuth.authenticate, function (req, res, next) {
     var response = {};
     if (req.params.sort == 'recent') {
         Ku.findAll({
@@ -102,26 +103,34 @@ POST a new ku. Body looks like:
     "lon": 28
 }
  */
-router.post('/compose', function (req, res, next) {
-    console.log(req.body);
-    var returnObject = {}
-    models.sequelize.transaction(function (t) {
-        return Ku.create({
-            content: req.body.ku,
-            lat: req.body.lat,
-            lon: req.body.lon
-        }, {transaction: t}).then(function (ku) {
-            returnObject.ku = ku.dataValues;
-            return Ku_user.create({
-                userId: req.body.userId,
-                kuId: ku.dataValues.id,
-                relationship: 0
-            }, {transaction: t});
-        });
-    }).then(function (result) {
-        res.json(returnObject);
-    }).catch(function (error) {
-        res.status(500).send(error);
+router.post('/compose', apiAuth.authenticate, function (req, res, next) {
+    var auth = req.get("authorization").split(' ')[1];
+    var auth_user = new Buffer(auth, 'base64').toString().split(':')[0];
+
+    User.findOne({where: {id: req.body.userId}}).then(function (user) {
+        if (user.dataValues.username == auth_user) {
+            var returnObject = {};
+            models.sequelize.transaction(function (t) {
+                return Ku.create({
+                    content: req.body.ku,
+                    lat: req.body.lat,
+                    lon: req.body.lon
+                }, {transaction: t}).then(function (ku) {
+                    returnObject.ku = ku.dataValues;
+                    return Ku_user.create({
+                        userId: req.body.userId,
+                        kuId: ku.dataValues.id,
+                        relationship: 0
+                    }, {transaction: t});
+                });
+            }).then(function (result) {
+                res.json(returnObject);
+            }).catch(function (error) {
+                res.status(500).send(error);
+            });
+        } else {
+            res.status(401).send('Unauthorized');
+        }
     });
 });
 
@@ -132,26 +141,35 @@ POST a newly favorited ku. Body looks like:
     "kuId": 15
 } 
 */
-router.post('/favorite', function (req, res, next) {
-    var whereClause = {
-            userId: req.body.userId,
-            kuId: req.body.kuId,
-            relationship: 1
-        }
-    Ku_user.findOrCreate({
-        where: whereClause
-    }).spread(function (ku_user, created) {
-        if (!created) {
-            Ku_user.destroy({
+router.post('/favorite', apiAuth.authenticate, function (req, res, next) {
+    var auth = req.get("authorization").split(' ')[1];
+    var auth_user = new Buffer(auth, 'base64').toString().split(':')[0];
+
+    User.findOne({where: {id: req.body.userId}}).then(function (user) {
+        if (user.dataValues.username == auth_user) {
+            var whereClause = {
+                    userId: req.body.userId,
+                    kuId: req.body.kuId,
+                    relationship: 1
+                }
+            Ku_user.findOrCreate({
                 where: whereClause
-            }).then(function () {
-                res.status(200).json({"Status": "Ku favorite removed"});
-            })
+            }).spread(function (ku_user, created) {
+                if (!created) {
+                    Ku_user.destroy({
+                        where: whereClause
+                    }).then(function () {
+                        res.status(200).json({"Status": "Ku favorite removed"});
+                    })
+                } else {
+                    res.status(200).json({"Status": "Ku favorite added"})
+                }
+            }).catch(function (error) {
+                res.status(500).send(error);
+            });
         } else {
-            res.status(200).json({"Status": "Ku favorite added"})
+            res.status(401).send('Unauthorized');
         }
-    }).catch(function (error) {
-        res.status(500).send(error);
     });
 });
 
@@ -162,34 +180,43 @@ POST an upvote to a ku. Body looks like:
     "kuId": 15
 } 
 */
-router.post('/upvote', function (req, res, next) {
-    var whereClause = {
-            userId: req.body.userId,
-            kuId: req.body.kuId,
-            relationship: 2
-        }
-    Ku_user.findOrCreate({
-        where: whereClause
-    }).spread(function (ku_user, created) {
-        if (!created) {
-            Ku_user.destroy({
+router.post('/upvote', apiAuth.authenticate, function (req, res, next) {
+    var auth = req.get("authorization").split(' ')[1];
+    var auth_user = new Buffer(auth, 'base64').toString().split(':')[0];
+
+    User.findOne({where: {id: req.body.userId}}).then(function (user) {
+        if (user.dataValues.username == auth_user) {
+            var whereClause = {
+                    userId: req.body.userId,
+                    kuId: req.body.kuId,
+                    relationship: 2
+                }
+            Ku_user.findOrCreate({
                 where: whereClause
-            }).then(function () {
-                Ku.findById(req.body.kuId).then(function (ku) {
-                    ku.decrement("upvotes");
-                    ku.decrement("karma");
-                    res.status(200).json({"Status": ku.dataValues.karma-1});
-                });
-            })
-        } else {
-            Ku.findById(req.body.kuId).then(function (ku) {
-                ku.increment("upvotes");
-                ku.increment("karma");
-                res.status(200).json({"Status": ku.dataValues.karma+1})
+            }).spread(function (ku_user, created) {
+                if (!created) {
+                    Ku_user.destroy({
+                        where: whereClause
+                    }).then(function () {
+                        Ku.findById(req.body.kuId).then(function (ku) {
+                            ku.decrement("upvotes");
+                            ku.decrement("karma");
+                            res.status(200).json({"Status": ku.dataValues.karma-1});
+                        });
+                    })
+                } else {
+                    Ku.findById(req.body.kuId).then(function (ku) {
+                        ku.increment("upvotes");
+                        ku.increment("karma");
+                        res.status(200).json({"Status": ku.dataValues.karma+1})
+                    });
+                }
+            }).catch(function (error) {
+                res.status(500).send(error);
             });
+        } else {
+            res.status(401).send('Unauthorized');
         }
-    }).catch(function (error) {
-        res.status(500).send(error);
     });
 });
 
@@ -200,52 +227,70 @@ POST a downvote to a ku. Body looks like:
     "kuId": 15
 } 
 */
-router.post('/downvote', function (req, res, next) {
-    var whereClause = {
-            userId: req.body.userId,
-            kuId: req.body.kuId,
-            relationship: 3
-        }
-    Ku_user.findOrCreate({
-        where: whereClause
-    }).spread(function (ku_user, created) {
-        if (!created) {
-            Ku_user.destroy({
+router.post('/downvote', apiAuth.authenticate, function (req, res, next) {
+    var auth = req.get("authorization").split(' ')[1];
+    var auth_user = new Buffer(auth, 'base64').toString().split(':')[0];
+
+    User.findOne({where: {id: req.body.userId}}).then(function (user) {
+        if (user.dataValues.username == auth_user) {
+            var whereClause = {
+                    userId: req.body.userId,
+                    kuId: req.body.kuId,
+                    relationship: 3
+                }
+            Ku_user.findOrCreate({
                 where: whereClause
-            }).then(function () {
-                Ku.findById(req.body.kuId).then(function (ku) {
-                    ku.decrement("downvotes");
-                    ku.increment("karma");
-                    res.status(200).json({"Status": ku.dataValues.karma+1});
-                })
-            })
+            }).spread(function (ku_user, created) {
+                if (!created) {
+                    Ku_user.destroy({
+                        where: whereClause
+                    }).then(function () {
+                        Ku.findById(req.body.kuId).then(function (ku) {
+                            ku.decrement("downvotes");
+                            ku.increment("karma");
+                            res.status(200).json({"Status": ku.dataValues.karma+1});
+                        })
+                    })
+                } else {
+                    Ku.findById(req.body.kuId).then(function (ku) {
+                        ku.increment("downvotes");
+                        ku.decrement("karma");
+                        res.status(200).json({"Status": ku.dataValues.karma-1})
+                    })
+                }
+            }).catch(function (error) {
+                res.status(500).send(error);
+            });
         } else {
-            Ku.findById(req.body.kuId).then(function (ku) {
-                ku.increment("downvotes");
-                ku.decrement("karma");
-                res.status(200).json({"Status": ku.dataValues.karma-1})
-            })
+            res.status(401).send('Unauthorized');
         }
-    }).catch(function (error) {
-        res.status(500).send(error);
     });
 });
 
 /*
 GET boolean check to see if user has already upvoted/downvoted a ku
 */
-router.get('/:id/:userId/:vote', function (req, res, next) {
-    var relationship = req.params.vote == 'upvote' ? 2 : 3;
-    Ku_user.findOne({
-        where: {
-            userId: req.params.userId,
-            kuId: req.params.id,
-            relationship: relationship
+router.get('/:id/:userId/:vote', apiAuth.authenticate, function (req, res, next) {
+    var auth = req.get("authorization").split(' ')[1];
+    var auth_user = new Buffer(auth, 'base64').toString().split(':')[0];
+
+    User.findOne({where: {id: req.params.userId}}).then(function (user) {
+        if (user.dataValues.username == auth_user) {
+            var relationship = req.params.vote == 'upvote' ? 2 : 3;
+            Ku_user.findOne({
+                where: {
+                    userId: req.params.userId,
+                    kuId: req.params.id,
+                    relationship: relationship
+                }
+            }).then(function (ku_user) {
+                var hasVote = ku_user !== null;
+                res.json({"status": hasVote});
+            });
+        } else {
+            res.status(401).send('Unauthorized');
         }
-    }).then(function (ku_user) {
-        var hasVote = ku_user !== null;
-        res.json({"status": hasVote});
-    })
-}) 
+    });
+});
 
 module.exports = router;
